Update bot.js
Updated the Spotify playlist loading process in the queue. Previously, it was passing the entire array without iterating through its elements to add them individually to the queue. This resulted in a nested array where elements couldn't be properly read, causing the endpoint to return null. This issue prevented the player from playing the songs in albums and playlists.